Refactor: Add comments to Dockerfile, rename download_sandboxes.py to upload_sandboxes_to_filetracker.py, and update the command’s functionality
